correct windows paths so slashes appear correctly on screen
There are references here to Windows path using normal path separators (\), but in markdown those must be escaped (when not preformatted text), otherwise they do not appear on screen. That was the case with a few of the references here. 

And note that one CANNOT do a global find/replace, because again the references in preformatted text should NOT be escaped (and there was in fact one reference with an escaped path separator, where it should not have been).