Update create.md
Hey guys,
I was not too sure if there is a style for writing but I noticed some small errors on the grammatical side. Also noticed some inconsistency when it came to the capitalization of the words "Environment Variables". My major change was the word "won't" as it's really unclear to me if it means wouldn't(conditional) or will not (future tense). I opted for will not but do correct it if it really ought to be the conditional,